requirement.txt
antlr4-python3-runtime==4.8
astroid==2.4.2
Brotli==1.0.7
click==7.1.2
cycler==0.10.0
dash==1.12.0
dash-core-components==1.10.0
dash-html-components==1.0.3
dash-renderer==1.4.1
dash-table==4.7.0
decorator==4.4.2
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
networkx==2.4
numpy==1.18.5
plotly==4.8.1
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
retrying==1.3.3
six==1.15.0
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
